+++ HISTORY.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+**Features and Improvements**
+
+* Added support for DictField - peterfarrell
+
 ## 0.6.0 (2020-01-03)
 
 **Features and Improvements**